--- src/App.tsx
@@ -1,25 +1,0 @@
-import React from 'react';
-
-import './App.css';
-
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { About, Contact, Home, Work, Error } from './pages';
-
-function App() {
-  return (
-    <>
-      <BrowserRouter>
-        <Routes>
-          <Route index element={<Home />} />
-          <Route path='/home' element={<Home />} />
-          <Route path='/work' element={<Work />} />
-          <Route path='/about' element={<About />} />
-          <Route path='/contact' element={<Contact />} />
-          <Route path='*' element={<Error />} />
-        </Routes>
-      </BrowserRouter>
-    </>
-  );
-}
-
-export default App;
