+++ src/index.tsx
@@ -3,13 +3,23 @@ import ReactDOM from 'react-dom/client';
 
 import './index.css';
 
-import App from './App';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { About, Contact, Home, Work, Error } from './pages';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <App />
+      <BrowserRouter>
+        <Routes>
+          <Route index element={<Home />} />
+          <Route path='/home' element={<Home />} />
+          <Route path='/work' element={<Work />} />
+          <Route path='/about' element={<About />} />
+          <Route path='/contact' element={<Contact />} />
+          <Route path='*' element={<Error />} />
+        </Routes>
+      </BrowserRouter>
   </React.StrictMode>
 );
