+++ src/pages/Contact.tsx
@@ -1,5 +1,5 @@
-import Nav from '../Nav';
-import Footer from '../Footer';
+import Nav from '../components/Nav';
+import Footer from '../components/Footer';
 
 export default function Contact() {
     return <>
